+++ lib/zip_filerange_crc.c
@@ -33,32 +33,30 @@
 
 
 #include <stdio.h>
-#include <stdlib.h>
 
 #include "zipint.h"
 
 
 int
 _zip_filerange_crc(zip_source_t *src, zip_uint64_t start, zip_uint64_t len, uLong *crcp, zip_error_t *error) {
-    Bytef *buf;
-    zip_int64_t n;
+    DEFINE_BYTE_ARRAY(buf, BUFSIZE);
 
-    if ((buf = (Bytef *)malloc(BUFSIZE)) == NULL) {
-	zip_error_set(error, ZIP_ER_MEMORY, 0);
-	return -1;
-    }
+    zip_int64_t n;
 
     *crcp = crc32(0L, Z_NULL, 0);
 
     if (start > ZIP_INT64_MAX) {
 	zip_error_set(error, ZIP_ER_SEEK, EFBIG);
-	free((void *)buf);
 	return -1;
     }
 
     if (zip_source_seek(src, (zip_int64_t)start, SEEK_SET) != 0) {
 	_zip_error_set_from_source(error, src);
-	free((void *)buf);
+	return -1;
+    }
+
+    if (!byte_array_init(buf, BUFSIZE)) {
+	zip_error_set(error, ZIP_ER_MEMORY, 0);
 	return -1;
     }
 
@@ -66,12 +64,12 @@ _zip_filerange_crc(zip_source_t *src, zip_uint64_t start, zip_uint64_t len, uLon
 	n = (zip_int64_t)(len > BUFSIZE ? BUFSIZE : len);
 	if ((n = zip_source_read(src, buf, (zip_uint64_t)n)) < 0) {
 	    _zip_error_set_from_source(error, src);
-	    free((void *)buf);
+	    byte_array_fini(buf);
 	    return -1;
 	}
 	if (n == 0) {
 	    zip_error_set(error, ZIP_ER_EOF, 0);
-	    free((void *)buf);
+	    byte_array_fini(buf);
 	    return -1;
 	}
 
@@ -80,6 +78,7 @@ _zip_filerange_crc(zip_source_t *src, zip_uint64_t start, zip_uint64_t len, uLon
 	len -= (zip_uint64_t)n;
     }
 
-    free((void *)buf);
+    byte_array_fini(buf);
+
     return 0;
 }